Add checkstyle and coding standard
