+++ build.gradle
@@ -8,6 +8,10 @@ repositories {
     mavenCentral()
 }
 
+checkstyle {
+    toolVersion = '10.2'
+}
+
 dependencies {
     testImplementation group: 'org.junit.jupiter', name: 'junit-jupiter-api', version: '5.10.0'
     testRuntimeOnly group: 'org.junit.jupiter', name: 'junit-jupiter-engine', version: '5.10.0'
