--- src/main/java/InputCommands/AddDeadlineCommand.java
@@ -1,64 +1,0 @@
-package InputCommands;
-
-import SnomExceptions.InvalidCommandDateFormatException;
-import SnomExceptions.InvalidCommandException;
-import SnomExceptions.InvalidCommandTaskDescException;
-import SnomTaskList.TaskList;
-
-import java.time.LocalDate;
-import java.time.format.DateTimeParseException;
-
-class AddDeadlineCommand extends Command {
-
-    @Override
-    public CmdType getType() {
-        return CmdType.DEADLINE;
-    }
-
-    protected AddDeadlineCommand(String desc) {
-        super(desc);
-    }
-
-    @Override
-    public String execute(TaskList lst) throws InvalidCommandException {
-
-        try {
-            //the format is as such: deadline name/end date
-            String detail = this.desc.toLowerCase().split("deadline ", 2)[1];
-            String name = detail.split("/", 2)[0].trim();
-            if (name.isEmpty()) {
-                throw new InvalidCommandTaskDescException();
-            }
-            String due_date = detail.split("/", 2)[1].trim(); //removes the whitspace from start and end of date
-            if (due_date.isEmpty()) {
-                throw new InvalidCommandTaskDescException();
-            }
-            LocalDate date = LocalDate.parse(due_date);
-
-
-
-            return name + "/" + date.toString();
-
-            //you can consider changing commands to return tasks instead;
-
-        } catch (ArrayIndexOutOfBoundsException e) {
-
-            throw new InvalidCommandTaskDescException();
-        } catch (DateTimeParseException e) {
-
-            throw new InvalidCommandDateFormatException();
-
-        }
-    }
-
-    public static void main(String[] args) {
-        try {
-
-            System.out.println(new AddDeadlineCommand("deadline /2000-01-01").execute(null));
-        } catch (Exception e) {
-            System.out.println(e.getMessage());
-        }
-    }
-
-
-}
