--- src/main/java/InputCommands/ByeCommand.java
@@ -1,26 +1,0 @@
-package InputCommands;
-
-import SnomExceptions.InvalidCommandException;
-import SnomTaskList.TaskList;
-
-
-public class ByeCommand extends Command {
-
-    private ByeCommand(String desc) {
-        super(desc);
-    }
-
-    protected ByeCommand() {
-        super(" ");
-    }
-
-    @Override
-    public CmdType getType() {
-        return CmdType.BYE;
-    }
-
-    @Override
-    public String execute(TaskList t) throws InvalidCommandException {
-        return "bye";
-    }
-}
