--- src/main/java/InputCommands/CmdType.java
@@ -1,5 +1,0 @@
-package InputCommands;
-
-public enum CmdType {
-    DEADLINE, EVENT, TODO, BYE, DELETE, LIST, MARK, UNMARK, FIND
-}
