--- src/main/java/SnomExceptions/InvalidCommandDateFormatException.java
@@ -1,8 +1,0 @@
-package SnomExceptions;
-
-
-public class InvalidCommandDateFormatException extends InvalidCommandException{
-    public InvalidCommandDateFormatException() {
-        super("Please ensure that your date(s) is of the format yyyy-mm-dd");
-    }
-}
