--- src/main/java/SnomExceptions/InvalidCommandDateValueException.java
@@ -1,8 +1,0 @@
-package SnomExceptions;
-
-public class InvalidCommandDateValueException extends InvalidCommandException{
-
-    public InvalidCommandDateValueException() {
-        super("Please make sure that your end date is after the start date");
-    }
-}
