--- src/main/java/SnomExceptions/InvalidCommandException.java
@@ -1,12 +1,0 @@
-package SnomExceptions;
-
-public class InvalidCommandException extends Exception{
-
-    public InvalidCommandException() {
-        super("Please enter a valid command");
-    }
-
-    public InvalidCommandException(String message) {
-        super(message);
-    }
-}
