--- src/main/java/SnomExceptions/InvalidCommandFormatException.java
@@ -1,8 +1,0 @@
-package SnomExceptions;
-
-public class InvalidCommandFormatException extends InvalidCommandException{
-
-    public InvalidCommandFormatException() {
-        super("Please make sure your command is of the valid format");
-    }
-}
