--- src/main/java/SnomExceptions/InvalidCommandIndexException.java
@@ -1,7 +1,0 @@
-package SnomExceptions;
-
-public class InvalidCommandIndexException extends InvalidCommandException{
-    public InvalidCommandIndexException() {
-        super("Please ensure that the index you entered is valid");
-    }
-}
