--- src/main/java/SnomExceptions/InvalidCommandTaskDescException.java
@@ -1,7 +1,0 @@
-package SnomExceptions;
-
-public class InvalidCommandTaskDescException extends InvalidCommandException{
-    public InvalidCommandTaskDescException() {
-        super("Please dont leave your task description or date blank");
-    }
-}
