--- src/main/java/SnomTask/Deadline.java
@@ -1,16 +1,0 @@
-package SnomTask;
-
-public class Deadline extends Task {
-
-    private String deadline;
-
-    public Deadline(String name, String deadline) {
-        super(name);
-        this.deadline = deadline;
-    }
-
-    @Override
-    public String toString() {
-        return "[D]" + super.toString() + " (by: " + this.deadline + ")";
-    }
-}
