--- src/main/java/SnomTask/Event.java
@@ -1,19 +1,0 @@
-package SnomTask;
-
-public class Event extends Task {
-
-    private String start;
-    private String end;
-
-    public Event(String name, String start, String end) {
-        super(name);
-        this.start = start;
-        this.end = end;
-    }
-
-
-    @Override
-    public String toString() {
-        return "[E]" + super.toString() + " from " + this.start + " to " + this.end;
-    }
-}
