--- src/main/java/SnomTask/Todo.java
@@ -1,14 +1,0 @@
-package SnomTask;
-
-public class Todo extends Task {
-
-    public Todo(String name) {
-        super(name);
-    }
-
-
-    @Override
-    public String toString() {
-        return "[T]" + super.toString();
-    }
-}
