+++ src/main/java/inputcommands/CmdType.java
@@ -1,0 +1,5 @@
+package inputcommands;
+
+public enum CmdType {
+    DEADLINE, EVENT, TODO, BYE, DELETE, LIST, MARK, UNMARK, FIND
+}
