+++ src/main/java/snomexceptions/InvalidCommandDateFormatException.java
@@ -1,0 +1,8 @@
+package snomexceptions;
+
+
+public class InvalidCommandDateFormatException extends InvalidCommandException {
+    public InvalidCommandDateFormatException() {
+        super("Please ensure that your date(s) is of the format yyyy-mm-dd");
+    }
+}
