+++ src/main/java/snomexceptions/InvalidCommandDateValueException.java
@@ -1,0 +1,8 @@
+package snomexceptions;
+
+public class InvalidCommandDateValueException extends InvalidCommandException {
+
+    public InvalidCommandDateValueException() {
+        super("Please make sure that your end date is after the start date");
+    }
+}
