+++ src/main/java/snomexceptions/InvalidCommandException.java
@@ -1,0 +1,12 @@
+package snomexceptions;
+
+public class InvalidCommandException extends Exception {
+
+    public InvalidCommandException() {
+        super("Please enter a valid command");
+    }
+
+    public InvalidCommandException(String message) {
+        super(message);
+    }
+}
