+++ src/main/java/snomexceptions/InvalidCommandFormatException.java
@@ -1,0 +1,8 @@
+package snomexceptions;
+
+public class InvalidCommandFormatException extends InvalidCommandException {
+
+    public InvalidCommandFormatException() {
+        super("Please make sure your command is of the valid format");
+    }
+}
