+++ src/main/java/snomexceptions/InvalidCommandIndexException.java
@@ -1,0 +1,7 @@
+package snomexceptions;
+
+public class InvalidCommandIndexException extends InvalidCommandException {
+    public InvalidCommandIndexException() {
+        super("Please ensure that the index you entered is valid");
+    }
+}
