+++ src/main/java/snomexceptions/InvalidCommandTaskDescException.java
@@ -1,0 +1,7 @@
+package snomexceptions;
+
+public class InvalidCommandTaskDescException extends InvalidCommandException {
+    public InvalidCommandTaskDescException() {
+        super("Please dont leave your task description or date blank");
+    }
+}
