+++ src/main/java/snomtask/Deadline.java
@@ -1,0 +1,16 @@
+package snomtask;
+
+public class Deadline extends Task {
+
+    private String deadline;
+
+    public Deadline(String name, String deadline) {
+        super(name);
+        this.deadline = deadline;
+    }
+
+    @Override
+    public String toString() {
+        return "[D]" + super.toString() + " (by: " + this.deadline + ")";
+    }
+}
