+++ src/main/java/snomtask/Todo.java
@@ -1,0 +1,14 @@
+package snomtask;
+
+public class Todo extends Task {
+
+    public Todo(String name) {
+        super(name);
+    }
+
+
+    @Override
+    public String toString() {
+        return "[T]" + super.toString();
+    }
+}
